+++ packages/meteor-components-ioc-plugin/HISTORY.md
@@ -1,3 +1,27 @@
+# 0.2.0
+
+**Jan. 1, 2016**
+
+- Drop support for depending on keys from the data context. As a workaround
+  you pass a key path to your component's `data()` method. This path request
+  is a reactive path. Behind the scenes the data is managed by an instance of
+  a [ReactiveObj](https://atmospherejs.com/xamfoo/reactive-obj).
+
+- Drop support for depending on `data` to get a reactive data context. As a
+  workaround call `currentData()` on the component.
+
+- Override the `data()` method to accept a key path. When calling this method
+  with a key path the request the request is reactive.
+
+- Add a `data(path)` helper to all components automatically that provides a
+  convenient helper to retrieve keys from the data context in a reactive
+  fashion. This allows individual data context paths to be reactive instead of
+  the entire data context.
+
+- Look for `inject` static method or array property and pass it as inject
+  parameter when installing services on their component's IOC container.
+
+
 # 0.1.2
 
 **Jan. 1, 2016**